requirements.txt
opencv-contrib-python
mysql-connector-python
numpy
pillow
pandas
openpyxl